# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 快灵感小程序 API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /ideas:
    get:
      summary: 获取灵感列表
      operationId: listIdeas
      tags: [Ideas]
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: 灵感列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListIdeasResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: 创建灵感
      operationId: createIdea
      tags: [Ideas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIdeaRequest"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateIdeaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  schemas:
    Idea:
      type: object
      required: [id, content, source, created_at]
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time
    CreateIdeaRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        source:
          type: string
          default: text
    CreateIdeaResponse:
      type: object
      required: [idea]
      properties:
        idea:
          $ref: "#/components/schemas/Idea"
    ListIdeasResponse:
      type: object
      required: [ideas, next_cursor, has_more]
      properties:
        ideas:
          type: array
          items:
            $ref: "#/components/schemas/Idea"
        next_cursor:
          type: string
        has_more:
          type: boolean
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"